mistral_chat/requirements.txt
mistralai
flask
python-dotenv
pytest
